Move Interpreter::unboxed to interpeter/utils
