+++ src/interpreter/utils.hpp
@@ -1,0 +1,7 @@
+#include "value_fwd.hpp"
+
+namespace Interpreter {
+
+Value* unboxed(Value* value);
+
+}
